Add handler response

+++ internal/data/db.go
@@ -30,7 +30,7 @@ func (d Database) AddTransfer(from, to string, value *big.Int) error {
 
 // struct will be moved
 type Transfer struct {
-	Id          int    `db:"id"`
+	Id          int64  `db:"id"`
 	FromAddress string `db:"from_address"`
 	ToAddress   string `db:"to_address"`
 	Value       string `db:"value"`
